1020 개인프로젝트

+++ ReFirstTask/src/co/ReTask/BoardApp.java
@@ -1,5 +1,6 @@
 package co.ReTask;
 
+import java.util.List;
 import java.util.Scanner;
 
 public class BoardApp {
@@ -11,38 +12,42 @@ public class BoardApp {
 		boolean run = true;
 		int j = 0;
 		int index = 0;
-
+		int pageNum=1, from=0, to=0;
 		while (run) {
 
-			// 글목록
+// 첫화면글목록
+			List<Board> list = dao.list();
+			int totalCnt = list.size();
+			int totalPage =(int) Math.ceil(totalCnt/5.0);
+			
+//			int pageNum=1, from=0, to=0;
+			from = (pageNum-1)*5; // 0
+			to = (pageNum*5)-1; // 4
 			System.out.println(String.format("%-5s", "Best") + dao.listRank().subString());
 			System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-//			for(Board bod : dao.list()) {
-//				System.out.println(bod.subString());
-//				System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-//			}
-
-			for (j = 0; j < dao.list().size(); j++) {
-				bod = dao.list().get(j);
+			for (j = 0; j < totalCnt; j++) {
+				if(j >= from && j <= to) {
+				bod = list.get(j);
 				System.out.println(String.format("%-5s", j + 1) + bod.subString());
-				System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+				
+				}
 			}
-
+			System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ"+pageNum+"/"+totalPage+"ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+			System.out.println("                      7.이전   8.다음");
 			System.out.println("====0.검색 1.글등록 2.상세보기 3.내가 쓴 글 4.종료 5.로그인 6.회원가입====");
 			int menu = scn.nextInt();
 			scn.nextLine();
 			if (menu == 1) {
-				// 글등록
+// 글등록
 				System.out.println(">>>>로그인 후 이용가능합니다<<<<");
 
 			} else if (menu == 2) {
-				// 상세보기
+// 상세보기
 				System.out.println("상세보기 번호>>");
 				int detaile2 = scn.nextInt();
 				scn.nextLine();
 				while (true) {
 					System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-//							System.out.println(dao.detail(detaile2));
 					System.out.println(detaile2 + dao.detail(dao.sublist().get(detaile2 - 1).getNum()).toString());
 					System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 					for (int e = 0; e < dao.comlist(dao.sublist().get(detaile2 - 1).getNum()).size(); e++) {
@@ -72,13 +77,13 @@ public class BoardApp {
 					}
 				}
 			} else if (menu == 3) {
-				// 내가 쓴 글 보기
+// 내가 쓴 글 보기
 				System.out.println(">>>>로그인 후 이용가능합니다<<<<");
 			} else if (menu == 4) {
-				// 종료
+// 종료
 				break;
 			} else if (menu == 5) {
-				// 로그인
+// 로그인
 				System.out.println("아이디를 입력해주세요>>");
 				String id = scn.nextLine();
 				System.out.println("비밀번호를 입력해주세요>>");
@@ -87,18 +92,23 @@ public class BoardApp {
 					System.out.println(">>>>로그인 되었습니다<<<<");
 
 					while (run) {
+						from = (pageNum-1)*5; // 0
+						to = (pageNum*5)-1; // 4
 						System.out.println(String.format("%-5s", "Best") + dao.listRank().subString());
 						System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-						for (j = 0; j < dao.list().size(); j++) {
-							bod = dao.list().get(j);
+						for (j = 0; j < totalCnt; j++) {
+							if(j >= from && j <= to) {
+							bod = list.get(j);
 							System.out.println(String.format("%-5s", j + 1) + bod.subString());
-							System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+							}
 						}
-						System.out.println("====0.검색 1.글등록 2.상세보기 3.내가 쓴 글 보기 4.종료====");
+						System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ"+pageNum+"/"+totalPage+"ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+						System.out.println("               7.이전   8.다음");
+						System.out.println("====0.검색 1.글등록 2.상세보기 3.내가 쓴 글 4.종료====");
 						int submenu = scn.nextInt();
 						scn.nextLine();
-						/* 글등록 */ if (submenu == 1) {
-							// 글등록
+						if (submenu == 1) {
+// 로그인 후 글등록
 							System.out.println("글제목>>");
 							String title = scn.nextLine();
 							System.out.println("글내용>>");
@@ -106,21 +116,17 @@ public class BoardApp {
 							Board boad = new Board(title, content, id);
 							dao.insert(boad);
 
-							/* 상세보기 */ } else if (submenu == 2) {
-							// 상세보기
+						} else if (submenu == 2) {
+// 로그인 후 상세보기
 							System.out.println("상세보기 번호>>");
 							int detaile = scn.nextInt();
 							scn.nextLine();
 							while (true) {
-								// 상세글
 								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 								System.out.println(
 										detaile + dao.detail(dao.sublist().get(detaile - 1).getNum()).toString());
 								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-								// 댓글목록
-//								for (Comment bod3 : dao.comlist(dao.sublist().get(detaile - 1).getNum())) {
-//									System.out.println("[" + bod3 + "]");
-//								}
+//댓글목록
 //댓글번호대로 정렬
 								for (int e = 0; e < dao.comlist(dao.sublist().get(detaile - 1).getNum()).size(); e++) {
 									Comment com = dao.comlist(dao.sublist().get(detaile - 1).getNum()).get(e);
@@ -129,7 +135,8 @@ public class BoardApp {
 								System.out.println("====1.수정 2.삭제 3.목록 4.댓글쓰기 5.댓글수정 6.댓글삭제====");
 								int detailmenu = scn.nextInt();
 								scn.nextLine();
-								/* 글수정 */ if (detailmenu == 1) {
+								if (detailmenu == 1) {
+//상세글 수정
 									System.out.println("제목 수정>>");
 									String retitle = scn.nextLine();
 									System.out.println("내용 수정>>");
@@ -140,7 +147,8 @@ public class BoardApp {
 									} else {
 										System.out.println(">>>>작성자만 수정할 수 있습니다<<<<");
 									}
-									/* 글삭제 */ } else if (detailmenu == 2) {
+								} else if (detailmenu == 2) {
+//상세글 삭제
 									if (dao.delete(dao.sublist().get(detaile - 1).getNum(), id) == 1) {
 										System.out.println(">>>>삭제되었습니다<<<<");
 										break;
@@ -148,13 +156,16 @@ public class BoardApp {
 										System.out.println(">>>>작성자만 삭제할 수 있습니다<<<<");
 									}
 
-									/* 글목록으로 돌아가기 */ } else if (detailmenu == 3) {
+								} else if (detailmenu == 3) {
+//상세글 목록으로 돌아가기
 									break;
-									/* 댓글쓰기 */ } else if (detailmenu == 4) {
+								} else if (detailmenu == 4) {
+//댓글쓰기
 									System.out.println("댓글 내용>>");
 									String comment = scn.nextLine();
 									dao.cominsert(id, comment, dao.sublist().get(detaile - 1).getNum());
-									/* 댓글수정 */ } else if (detailmenu == 5) {
+								} else if (detailmenu == 5) {
+//댓글수정
 									System.out.println("수정할 댓글번호>>");
 									int comnum = scn.nextInt();
 									scn.nextLine();
@@ -166,7 +177,8 @@ public class BoardApp {
 										System.out.println(">>>>작성자가 아닙니다<<<<");
 									}
 
-									/* 댓글삭제 */ } else if (detailmenu == 6) {
+								} else if (detailmenu == 6) {
+//댓글삭제
 									System.out.println("삭제할 댓글번호>>");
 									int comnum2 = scn.nextInt();
 									scn.nextLine();
@@ -178,17 +190,15 @@ public class BoardApp {
 
 							}
 						} else if (submenu == 3) {
-							// 내가 쓴 글 보기
-//							for (Board bod2 : dao.mylist(id)) {
-//								System.out.println(bod2.subString());
-//							}
+//로그인 후 내가 쓴 글 보기
 							boolean run7 = true;
-							while(run7) {
-							for(int p =0; p<dao.mylist(id).size();p++) {
-								bod3 = dao.mylist(id).get(p);
-								System.out.println(dao.mylistindex(id).get(p) + bod3.subString());
+							while (run7) {
 								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-							}
+								for (int p = 0; p < dao.mylist(id).size(); p++) {
+									bod3 = dao.mylist(id).get(p);
+									System.out.println(dao.mylistindex(id).get(p) + bod3.subString());
+								}
+								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 								System.out.println("====1.상세보기 2.목록====");
 								int menu6 = scn.nextInt();
 								scn.nextLine();
@@ -198,13 +208,13 @@ public class BoardApp {
 									scn.nextLine();
 									boolean run4 = true;
 									while (run4) {
-										// 상세글
-										System.out.println(
-												"ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-										System.out.println(detailenum0 + dao
-												.detail(dao.sublist().get(detailenum0 - 1).getNum()).toString());
-										System.out.println(
-												"ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+//내가 쓴 글에서 상세글
+										System.out
+												.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
+										System.out.println(detailenum0
+												+ dao.detail(dao.sublist().get(detailenum0 - 1).getNum()).toString());
+										System.out
+												.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 										// 댓글번호대로 정렬
 										for (int e = 0; e < dao.comlist(dao.sublist().get(detailenum0 - 1).getNum())
 												.size(); e++) {
@@ -215,7 +225,8 @@ public class BoardApp {
 										System.out.println("====1.수정 2.삭제 3.목록 4.댓글쓰기 5.댓글수정 6.댓글삭제====");
 										int detailmenu = scn.nextInt();
 										scn.nextLine();
-										/* 글수정 */ if (detailmenu == 1) {
+										if (detailmenu == 1) {
+//내가 쓴 글에서 글수정
 											System.out.println("제목 수정>>");
 											String retitle = scn.nextLine();
 											System.out.println("내용 수정>>");
@@ -226,7 +237,8 @@ public class BoardApp {
 											} else {
 												System.out.println(">>>>작성자만 수정할 수 있습니다<<<<");
 											}
-											/* 글삭제 */ } else if (detailmenu == 2) {
+										} else if (detailmenu == 2) {
+//내가 쓴 글에서 글삭제
 											if (dao.delete(dao.sublist().get(detailenum0 - 1).getNum(), id) == 1) {
 												System.out.println(">>>>삭제되었습니다<<<<");
 												break;
@@ -234,20 +246,22 @@ public class BoardApp {
 												System.out.println(">>>>작성자만 삭제할 수 있습니다<<<<");
 											}
 
-											/* 글목록으로 돌아가기 */ } else if (detailmenu == 3) {
+										} else if (detailmenu == 3) {
+//내가 쓴 글에서 목록으로 돌아가기
 											run7 = false;
 											run4 = false;
-											/* 댓글쓰기 */ } else if (detailmenu == 4) {
+										} else if (detailmenu == 4) {
+//내가 쓴 글에서 댓글쓰기
 											System.out.println("댓글 내용>>");
 											String comment = scn.nextLine();
 											dao.cominsert(id, comment, dao.sublist().get(detailenum0 - 1).getNum());
-											/* 댓글수정 */ } else if (detailmenu == 5) {
+										} else if (detailmenu == 5) {
+//내가 쓴 글에서 댓글수정
 											System.out.println("수정할 댓글번호>>");
 											int comnum = scn.nextInt();
 											scn.nextLine();
 											System.out.println("수정할 내용>>");
 											String comcon = scn.nextLine();
-											// dao.comupdate(comcon, comnum);
 											if (dao.comupdate(comcon,
 													dao.comlist2(dao.sublist().get(detailenum0 - 1).getNum())
 															.get(comnum - 1).getCnum(),
@@ -255,7 +269,8 @@ public class BoardApp {
 												System.out.println(">>>>작성자가 아닙니다<<<<");
 											}
 
-											/* 댓글삭제 */ } else if (detailmenu == 6) {
+										} else if (detailmenu == 6) {
+//내가 쓴 글에서 댓글삭제
 											System.out.println("삭제할 댓글번호>>");
 											int comnum2 = scn.nextInt();
 											scn.nextLine();
@@ -272,25 +287,23 @@ public class BoardApp {
 								}
 
 							}
-						
+
 /////////////////////////////////////
 						} else if (submenu == 4) {
-							// 종료
+// 로그인 후 종료
 							run = false;
 							System.out.println("종료되었습니다");
 
 						} else if (submenu == 0) {
-							// 검색
+// 로그인 후 검색
 							boolean run3 = true;
 							while (run3) {
 								System.out.println("검색>>");
 								String contentSearch = scn.nextLine();
-//								for (Board bod2 : dao.search(contentSearch)) {
-//									System.out.println(dao.listindex(contentSearch).get(index++) + bod2.subString());
 								for (int o = 0; o < dao.search(contentSearch).size(); o++) {
 									System.out.println(dao.listindex(contentSearch).get(o)
 											+ dao.search(contentSearch).get(o).subString());
-
+									}
 									System.out.println("====1.상세보기 2.목록====");
 									int menu3 = scn.nextInt();
 									scn.nextLine();
@@ -300,7 +313,7 @@ public class BoardApp {
 										scn.nextLine();
 										boolean run4 = true;
 										while (run4) {
-											// 상세글
+// 로그인 후 검색에서 상세글
 											System.out.println(
 													"ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 											System.out.println(detailenum + dao
@@ -317,7 +330,8 @@ public class BoardApp {
 											System.out.println("====1.수정 2.삭제 3.목록 4.댓글쓰기 5.댓글수정 6.댓글삭제====");
 											int detailmenu = scn.nextInt();
 											scn.nextLine();
-											/* 글수정 */ if (detailmenu == 1) {
+											if (detailmenu == 1) {
+// 로그인 후 검색에서 상세글 수정
 												System.out.println("제목 수정>>");
 												String retitle = scn.nextLine();
 												System.out.println("내용 수정>>");
@@ -328,7 +342,8 @@ public class BoardApp {
 												} else {
 													System.out.println(">>>>작성자만 수정할 수 있습니다<<<<");
 												}
-												/* 글삭제 */ } else if (detailmenu == 2) {
+											} else if (detailmenu == 2) {
+// 로그인 후 검색에서 상세글 삭제
 												if (dao.delete(dao.sublist().get(detailenum - 1).getNum(), id) == 1) {
 													System.out.println(">>>>삭제되었습니다<<<<");
 													break;
@@ -336,20 +351,22 @@ public class BoardApp {
 													System.out.println(">>>>작성자만 삭제할 수 있습니다<<<<");
 												}
 
-												/* 글목록으로 돌아가기 */ } else if (detailmenu == 3) {
+											} else if (detailmenu == 3) {
+// 로그인 후 검색에서 상세글 목록으로 돌아가기
 												run3 = false;
 												run4 = false;
-												/* 댓글쓰기 */ } else if (detailmenu == 4) {
+											} else if (detailmenu == 4) {
+// 로그인 후 검색에서 상세글 댓글쓰기
 												System.out.println("댓글 내용>>");
 												String comment = scn.nextLine();
 												dao.cominsert(id, comment, dao.sublist().get(detailenum - 1).getNum());
-												/* 댓글수정 */ } else if (detailmenu == 5) {
+											} else if (detailmenu == 5) {
+// 로그인 후 검색에서 상세글 댓글수정
 												System.out.println("수정할 댓글번호>>");
 												int comnum = scn.nextInt();
 												scn.nextLine();
 												System.out.println("수정할 내용>>");
 												String comcon = scn.nextLine();
-												// dao.comupdate(comcon, comnum);
 												if (dao.comupdate(comcon,
 														dao.comlist2(dao.sublist().get(detailenum - 1).getNum())
 																.get(comnum - 1).getCnum(),
@@ -357,7 +374,8 @@ public class BoardApp {
 													System.out.println(">>>>작성자가 아닙니다<<<<");
 												}
 
-												/* 댓글삭제 */ } else if (detailmenu == 6) {
+											} else if (detailmenu == 6) {
+// 로그인 후 검색에서 상세글 댓글삭제
 												System.out.println("삭제할 댓글번호>>");
 												int comnum2 = scn.nextInt();
 												scn.nextLine();
@@ -372,8 +390,18 @@ public class BoardApp {
 									} else if (menu3 == 2) {
 										run3 = false;
 									}
-
-								}
+							}
+						}else if(submenu ==7) {
+							if(pageNum > 1) {
+								pageNum = pageNum-1;
+							}else {
+								System.out.println("첫번째 페이지 입니다");
+							}
+						}else if(submenu ==8) {
+							if(pageNum < totalPage) {
+								pageNum = pageNum+1;
+							}else{
+								System.out.println("마지막 페이지 입니다");
 							}
 						}
 					} // end of while
@@ -382,7 +410,7 @@ public class BoardApp {
 				}
 
 			} else if (menu == 6) {
-				// 회원가입
+// 회원가입
 				System.out.println("아이디를 입력해주세요>>");
 				String id = scn.nextLine();
 				System.out.println("비밀번호를 입력해주세요>>");
@@ -391,17 +419,16 @@ public class BoardApp {
 				String name = scn.nextLine();
 				dao.memberShip(id, pw, name);
 
-				/* 검색 */ } else if (menu == 0) {
+			} else if (menu == 0) {
+//로그인 전 검색
 				boolean run1 = true;
 				while (run1) {
 					System.out.println("검색>>");
 					String contentSearch = scn.nextLine();
-//						for (Board bod2 : dao.search(contentSearch)) {
-//							System.out.println(dao.listindex(contentSearch).get(index++) + bod2.subString());
 					for (int o = 0; o < dao.search(contentSearch).size(); o++) {
 						System.out.println(
 								dao.listindex(contentSearch).get(o) + dao.search(contentSearch).get(o).subString());
-
+						}
 						System.out.println("====1.상세보기 2.목록====");
 						int menu5 = scn.nextInt();
 						scn.nextLine();
@@ -411,12 +438,12 @@ public class BoardApp {
 							scn.nextLine();
 							boolean run5 = true;
 							while (run5) {
-								// 상세글
+// 로그인 전 상세글
 								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
 								System.out.println(
 										detailenum + dao.detail(dao.sublist().get(detailenum - 1).getNum()).toString());
 								System.out.println("ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ");
-								// 댓글번호대로 정렬
+// 댓글번호대로 정렬
 								for (int e = 0; e < dao.comlist(dao.sublist().get(detailenum - 1).getNum())
 										.size(); e++) {
 									Comment com = dao.comlist(dao.sublist().get(detailenum - 1).getNum()).get(e);
@@ -425,24 +452,29 @@ public class BoardApp {
 								System.out.println("====1.수정 2.삭제 3.목록 4.댓글쓰기 5.댓글수정 6.댓글삭제====");
 								int detailmenu2 = scn.nextInt();
 								scn.nextLine();
-								/* 글수정 */ if (detailmenu2 == 1) {
+								if (detailmenu2 == 1) {
+//로그인 전 글수정
 									System.out.println(">>>>로그인 후 이용가능<<<<");
 									break;
-									/* 글삭제 */ } else if (detailmenu2 == 2) {
-
+								} else if (detailmenu2 == 2) {
+//로그인 전 글삭제
 									System.out.println(">>>>로그인 후 이용가능<<<<");
 									break;
 
-									/* 글목록으로 돌아가기 */ } else if (detailmenu2 == 3) {
+								} else if (detailmenu2 == 3) {
+//로그인 전 글 목록으로 돌아가기
 									run5 = false;
-									/* 댓글쓰기 */ } else if (detailmenu2 == 4) {
+								} else if (detailmenu2 == 4) {
+//로그인 전 댓글쓰기
 									System.out.println(">>>>로그인 후 이용가능<<<<");
 									break;
-									/* 댓글수정 */ } else if (detailmenu2 == 5) {
+								} else if (detailmenu2 == 5) {
+//로그인 전 댓글수정
 									System.out.println(">>>>로그인 후 이용가능<<<<");
 									break;
 
-									/* 댓글삭제 */ } else if (detailmenu2 == 6) {
+								} else if (detailmenu2 == 6) {
+//로그인 전 댓글삭제
 									System.out.println(">>>>로그인 후 이용가능<<<<");
 									break;
 
@@ -454,8 +486,18 @@ public class BoardApp {
 							run1 = false;
 							break;
 						}
-
-					}
+				}
+			}else if(menu == 7){
+				if(pageNum > 1) {
+					pageNum = pageNum-1;
+				}else {
+					System.out.println("첫번째 페이지 입니다");
+				}
+			}else if(menu == 8) {
+				if(pageNum < totalPage) {
+					pageNum = pageNum+1;
+				}else{
+					System.out.println("마지막 페이지 입니다");
 				}
 			}
 		} // end of while
